+++ app/interdiction-location/page.tsx
@@ -14,6 +14,7 @@ import Image from 'next/image'
 import Link from 'next/link'
 import { HeaderWrapper } from '../LandingUI'
 import { Calendrier } from './UI'
+import { timeIsRunningOut, timeIsRunningOut } from '@/components/timeUtils'
 
 export const description = `Les passoires thermiques seront progressivement interdites à la location. Propriétaire bailleur, découvrez vos aides à la rénovation énergétique.`
 
@@ -167,30 +168,5 @@ export default function Page() {
         </Wrapper>
       </PageBlock>
     </main>
-  )
-}
-
-const timeIsRunningOut = (to) => {
-  const toDate = new Date(Date.parse(to))
-  const now = new Date()
-  const months = monthDiff(now, toDate)
-
-  if (months < 12) return months + ' mois'
-  if (months < 12 + 3) return '1 an'
-  if (months < 12 + 8) return '1 an et demi'
-  if (months < 12 + 12) return 'moins de 2 ans'
-
-  const years = Math.floor(months / 12),
-    reminder = months % 12
-  if (years > 8) return years + ' ans'
-  if (reminder > 6) return `moins de ${years + 1} ans`
-  return `${years} ans`
-}
-
-function monthDiff(dateFrom, dateTo) {
-  return (
-    dateTo.getMonth() -
-    dateFrom.getMonth() +
-    12 * (dateTo.getFullYear() - dateFrom.getFullYear())
   )
 }
